requirements.txt
fake_useragent
requests
fire==0.1.0
pillow
itchat